Create README.md
I developed a passion for data analysis during my scientific work. Possession of skills in python and excel allow me to confirm hypotheses not only in scientific, but also in marketing research. I diligently explore data, apply statistical tools, and use data visualization to demonstrate project findings. I am sure that in our time, marketers and scientists cannot do without an in-depth knowledge of programming. That is why I spend a lot of time learning new libraries.

In 2022 he became a candidate of political sciences. My work skills can be seen in scientific articles and commercial projects. In my projects, I study a lot of attention to the social, cultural and political context of the research. Knowledge in the field of philosophy and political science allows me to make interesting hypotheses.